requirements.txt
asgiref==3.7.2
certifi==2023.5.7
distlib==0.3.6
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.2
pipenv==2023.7.11
platformdirs==3.8.1
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.23.1
virtualenv-clone==0.5.7
gunicorn==21.2.0